[Fix] export 시 인스턴스가 아닌 클래스로 변경 (#15)
- typeDi Container 사용 함수 생성, get & setDependency
