+++ src/domain/repositories/UserRepository.js
@@ -1,10 +1,6 @@
 /* eslint-disable no-unused-vars */
 /* eslint-disable class-methods-use-this */
 export default class UserRepository {
-  constructor(Entity) {
-    this.entity = Entity;
-  }
-
   findByEmail(email) {
     throw new Error('ERR_METHOD_SHOULD_OVERRIDE');
   }
